requirements.txt
aiogram==3.4.1
aiohttp
python-dotenv
requests
aiofiles
